backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-ollama
mysql-connector-python
python-dotenv
requests
